% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2cross-data.R
\docType{data}
\name{f2cross}
\alias{f2cross}
\title{Example experimental cross data}
\format{An object of class \code{"cross"}. See
\code{\link[qtl]{read.cross}} in the R/qtl package for details.}
\usage{
data(f2cross)
}
\description{
Simulated experimental cross data with some sample mix-ups. The
only phenotype is an individual ID. There are 100 individuals
genotyped at 1000 markers on 19 autosomes.
}
\examples{
library(qtl)
data(f2cross)
summary(f2cross)
}
\seealso{
\code{\link{expr1}}, \code{\link{expr2}}, \code{\link{genepos}}, \code{\link{pmap}}
}
\keyword{datasets}
