\name{lineup-internal}
\alias{print.summary.lineupdist}
\alias{print.lineupdist}

\title{Internal lineup functions}

\description{
  Internal lineup functions.
}

\usage{
\method{print}{summary.lineupdist}(x, \dots)
\method{print}{lineupdist}(x, \dots)
}

\details{
  These are generally not to be called directly by the user.
}

\author{Karl W Broman \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
