% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{lineupversion}
\alias{lineupversion}
\title{Installed version of R/lineup}
\usage{
lineupversion()
}
\value{
A character string with the version number of the currently
installed version of R/lineup.
}
\description{
Print the version number of the currently installed version of R/lineup.
}
\examples{
lineupversion()
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\keyword{print}

