% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap-data.R
\docType{data}
\name{pmap}
\alias{pmap}
\title{Physical map of markers}
\format{A list of vectors, each containing the locations of markers in Mbp.
(Technically, an object of class \code{"map"}.)}
\usage{
data(pmap)
}
\description{
Physical map (Mbp positions) of the markers in \code{\link{f2cross}}
}
\examples{
data(pmap)
summary(pmap)
plot(pmap)
}
\seealso{
\code{\link{expr1}}, \code{\link{expr2}}, \code{\link{f2cross}}, \code{\link{genepos}}
}
\keyword{datasets}
