% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-package.R
\docType{package}
\name{lineup-package}
\alias{lineup}
\alias{lineup-package}
\title{lineup: Lining Up Two Sets of Measurements}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for detecting and correcting sample mix-ups between two sets of measurements, such as between gene expression data on two tissues. Broman et al. (2015) \doi{10.1534/g3.115.019778}.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/lineup/lineup.html}{R/lineup User Guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/lineup}
  \item Report bugs at \url{https://github.com/kbroman/lineup/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
