\name{lineupversion}
\alias{lineupversion}

\title{Installed version of R/lineup}

\description{
  Print the version number of the currently installed version of R/lineup.
}

\usage{lineupversion()}

%\arguments{
%}

\value{
  A character string with the version number of the currently installed
  version of R/lineup.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\examples{
  lineupversion()
}


\keyword{print}
